+++ lib/collaboration_web/views/idea_view.ex
@@ -64,7 +64,8 @@ defmodule CollaborationWeb.IdeaView do
     {rating, all_raters}
   end
 
-  def render("idea.json", %{idea: i, user: u}) do
+  def render("idea.json", %{ idea: i, user: u, bot_comments: bot_comments }) do
+
 
     inserted_at = if u.condition == 0 || i.user_id == u.id,
       do: i.inserted_at,
@@ -76,20 +77,29 @@ defmodule CollaborationWeb.IdeaView do
 
     { rating, raters } = calc_rating(i.fake_rating, i.fake_raters, my_rating)
 
-    comments = if Ecto.assoc_loaded?(i.comments) do
+    comments =
+      # add relevant bot comments
       i.comments
+      |> Enum.filter(fn c -> Enum.member?(u.comments, c.id) end)
+      |> Enum.map(fn c -> Enum.find_index(u.comments, fn x -> x == c.id end) end)
+      |> Enum.map(fn i -> Enum.fetch!(bot_comments, i) end)
+      # and existing comments for current idea
+      |> Enum.concat(i.comments)
       |> View.render_many(CommentView, "comment.json", user: u)
       |> Enum.sort_by(fn(c) -> c.inserted_at end)
-    else
-      []
-    end
 
-    # if user was not preloaded check if id matches current user
-    user = cond do
-      Ecto.assoc_loaded?(i.user) -> i.user.name
-      i.user_id == u.id -> u.name
-      true -> "Unknown"
-    end
+    reload_in =
+      comments
+      |> Enum.map(fn c -> c.inserted_at end)
+      |> Enum.filter(fn t -> remaining(t) > 0 end)
+      |> List.first()
+
+    reload_in =
+      case reload_in do
+        nil -> nil
+        time -> remaining(time) * 1000
+      end
+
 
     %{
       id: i.id,
@@ -99,8 +109,9 @@ defmodule CollaborationWeb.IdeaView do
       rating: if is_nil(rating) do nil else Float.round(rating, 1) end,
       raters: raters,
       text: i.text,
+      reload_in: reload_in,
       user_id: i.user_id,
-      user: user
+      user: i.user.name
     }
   end